Sync User Workspace Cache (pjs182) (#75)
* Remove workspace roles when deactivating workspace

Added a function as a background to handle this. Function is in draft, has not been tested yet

* Finished draft for restoring workspace roles

* Added exception handling

* Perform a background task depending on the status

* Testing and fixing feature

Tested feature in local instance, working as expected according to expected behavior criterial

* Formatted file

Via `nix fmt`. Other changes were suggested, but for the sake of being compact I am leaving them out.

* Formatted other files

It seems that the formatter portion of Github Actions is still failing, so this commit is just formatting other files, according to the `nix fmt` command.

* Modified `get_thread_list` to use `user_id`

Implementing changes required for #76

* Changed type indicator and "all" indicator

Works properly now for teachers and students

* Updated file formatting

* Modifying existing set_user_role_with_student_id

Changing this endpoint to allow admins to only see the ID of a user being added for the sake of anonymity. There are *many* more instances of `student_id` being used within the API, but this is the only one which immediately is apparent, outside of the tables, in the frontend

* Formatting changes

* remove unused endpoint and update type hints

* update access map for set_user_role_with_user_id to restrict teacher access

* remove unused chat generator functions and update model name for ChatAnthropic

* update DynamoDB table name based on environment configuration

---------

Co-authored-by: Ruihuang Yang <rxy216@case.edu>